feat(cli): finish deploy command
