+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@semoss/cli",
-    "version": "1.0.0-beta.13",
+    "version": "1.0.0-beta.15",
     "type": "module",
     "types": "dist/index.d.ts",
     "files": [
